+++ examples/8_simple_loader.py
@@ -3,7 +3,7 @@ from typing import List, Optional
 from pydantic import BaseModel
 from pydantic_resolve import Resolver, mapper, LoaderDepend
 
-# loader functions
+# define loader functions
 async def friends_batch_load_fn(names):
     mock_db = {
         'tangkikodo': ['tom', 'jerry'],
@@ -11,11 +11,7 @@ async def friends_batch_load_fn(names):
         'trump': ['sam', 'jim'],
         'sally': ['sindy', 'lydia'],
     }
-    result = []
-    for name in names:
-        friends = mock_db.get(name, [])
-        result.append(friends)
-    return result
+    return [mock_db.get(name, []) for name in names]
 
 async def contact_batch_load_fn(names):
     mock_db = {
@@ -23,15 +19,12 @@ async def contact_batch_load_fn(names):
         'jim': 600, 'sindy': 700, 'lydia': 800, 'tangkikodo': 900, 'john': 1000,
         'trump': 1200, 'sally': 1300,
     }
-    result = []
-    for name in names:
-        contact = mock_db.get(name, None)
-        result.append(contact)
-    return result
+    return [mock_db.get(name, None) for name in names]
 
-# schemas
+# define schemas
 class Contact(BaseModel):
     number: Optional[int]
+
 class Friend(BaseModel):
     name: str
 
@@ -39,6 +32,7 @@ class Friend(BaseModel):
     @mapper(lambda n: Contact(number=n))
     def resolve_contact(self, loader=LoaderDepend(contact_batch_load_fn)):
         return loader.load(self.name)
+
 class User(BaseModel):
     name: str
     age: int
@@ -59,13 +53,14 @@ class User(BaseModel):
 
 class Root(BaseModel):
     users: List[User] = []
+    @mapper(lambda items: [User(**item) for item in items])
     def resolve_users(self):
         return [
-            User(name="tangkikodo", age=19), 
-            User(name='john', age=21), 
-            # User(name='trump', age=59), 
-            # User(name='sally', age=21), 
-            # User(name='some one', age=0)
+            {"name": "tangkikodo", "age": 19}, 
+            {"name": "john", "age": 20}, 
+            {"name": "trump", "age": 21}, 
+            {"name": "sally", "age": 22}, 
+            {"name": "no man", "age": 23}, 
         ]
 
 async def main():