+++ examples/fastapi_demo/loader.py
@@ -1,7 +1,4 @@
-from collections import defaultdict
-from mimetypes import common_types
 from aiodataloader import DataLoader
-import fastapi_demo.schema as sc
 import fastapi_demo.model as sm
 import fastapi_demo.db as db
 from sqlalchemy import select
@@ -16,15 +13,11 @@ class FeedbackLoader(DataLoader):
                 .where(sm.Feedback.private==self.private)  # <-------- global filter
                 .where(sm.Feedback.comment_id.in_(comment_ids)))
             rows = res.scalars().all()
-            items = [sc.FeedbackSchema.from_orm(row) for row in rows]
-
-            return build_list(items, comment_ids, lambda x: x.comment_id)
+            return build_list(rows, comment_ids, lambda x: x.comment_id)
 
 class CommentLoader(DataLoader):
     async def batch_load_fn(self, task_ids):
         async with db.async_session() as session:
             res = await session.execute(select(sm.Comment).where(sm.Comment.task_id.in_(task_ids)))
             rows = res.scalars().all()
-            items = [sc.CommentSchema.from_orm(row) for row in rows]
-
-            return build_list(items, task_ids, lambda x: x.task_id)
+            return build_list(rows, task_ids, lambda x: x.task_id)