+++ examples/fastapi_demo/schema.py
@@ -1,7 +1,7 @@
 from asyncio import Future
 from pydantic import BaseModel
 from typing import List
-from pydantic_resolve import LoaderDepend
+from pydantic_resolve import LoaderDepend, mapper
 import fastapi_demo.loader as ld
 
 
@@ -20,6 +20,7 @@ class CommentSchema(BaseModel):
     content: str
 
     feedbacks: List[FeedbackSchema] = [] 
+    @mapper(lambda items: [FeedbackSchema.from_orm(i) for i in items])
     def resolve_feedbacks(self, feedback_loader=LoaderDepend(ld.FeedbackLoader)) -> Future[List[FeedbackSchema]]:
         return feedback_loader.load(self.id)
 
@@ -31,6 +32,7 @@ class TaskSchema(BaseModel):
     name: str
 
     comments: List[CommentSchema] = [] 
+    @mapper(lambda items: [CommentSchema.from_orm(i) for i in items])
     def resolve_comments(self, comment_loader=LoaderDepend(ld.CommentLoader)) -> Future[List[CommentSchema]]:
         return comment_loader.load(self.id)