+++ changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## v1.1.1 (2023.6.17)
+
+- extend @mapper decorator with target class option, this will call auto_mapping function inside.
+
+```python
+comments: List[CommentSchema]  = []
+@mapper(CommentSchema)
+def resolve_comments(self, loader=LoaderDepend(CommentLoader)):
+    return loader.load(self.id)
+```
+
 ## v1.1.0 (2023.6.16)
 
 - add @mapper decorator, to enable custom data transform